Move to gadgets and wrap it as a lib untility
