+++ halo2_gadgets/src/lib.rs
@@ -30,3 +30,4 @@ pub mod poseidon;
 pub mod sha256;
 pub mod sinsemilla;
 pub mod utilities;
+pub mod poly_eval;
